[34] Wrapped experimental combobox into a serviceComboBox component
